backend/requirements.txt
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
pypdf
sentence-transformers
python-dotenv
fastapi
uvicorn[standard]
pydantic>=2.0
httpx
pytest
tqdm
